requirements.txt
Django=5.0.7
djangorestframework=3.15.2
async-lru=2.0.4
async-timeout=4.0.3